Create well formed HTML from build parser
The html from the build parser was working html, but was not well formed which meant that folding in editors was not working as expected.
The parser has been updated to ensure that all end tags are included. The main culprits were missing </p>, </td> tags, but also some others.
Some html comments have been added to make reading the output easier.