Format diaply of openers to have consistent width
When parsing build files each round of an opener had its width defined by the text associated with it. This led to variable width columns that was both difficult to read and unsightly.
This commit fixes the width of each round at 250px and centres both the text and the images within this. The height of the cell is now adjusted to meet the text needs rather than the width.