Have build parser default to parsing all builds
The build parser required the specific build to process be passed as a mandatory argument. This was effective during core development, but now changes to the output from the parser are more likely to be required across all builds. In order to simplify this the build id to process is an optional argument with the default being to process all builds in the specified data file.