Make normal level template repeatable
The template did not have a comma at the end of the standard level block. This block will be copy/pasted multiple times so it makes more sense for it to have the comma at the end and require that it be removed for the final level rather than requiring it to be added every level befpre then.