Update template creation script to allow intermediate testing
The script to construct the build template json file created all parts, but made it awkward to test the build structure iteratively.
This commit updates the script to create an additional shortened template file only containing the core elements of a build id is passed in. This allows the shortened template be used one level at a time to allow visualisation of the hand. The full template is still created so may be used to copy in additional blocks as required.