Remove duplicate level deck cards
In cases where multiple cards were changed during a level pick and one of the cards was from a previous level it would be duplicated in the full deck available for that level and subsequent ones. This commit fixes that by only adding cards to the full deck if they are not already present.