Allow more detailed class description
The description tag was an array that allowed a set of paragraphs be defined, which for many classes this was sufficent. Some classes however benefit from a more structured and complex description. In order to allow this the original tag has been renamed to "overview" and a new "description" tag added that allows for a heading and a number of div elements which act as containers for any html elements, greatly increasing the available formatting options.

All the existing class data files have been updated to use the new tag name and the html pages recreated. In some cases the original pages had tabs rather than spaces as the leading whitespaces this recreation normalises them all to be consistent.