Allow more detailed build description
The existing build json format allows for a basic overview that contains 3 sections with fixed text headings for two of them. This was created as it fit the format of the Frosthaven build guides, but is very limiting in its structure. This commit allows for a new json structure which can define multiple sections each with their own html heading level and text and free context content that is not tied to paragraphs.